requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.7.3
requests==2.32.3
werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1 # For loading .env file during local development

# Core ML/AI Libraries - Pinned versions for consistency and deployment
torch==2.3.1 # Or the exact version you installed with CUDA locally
torchvision==0.18.1 # Or the exact version you installed with CUDA locally
numpy==1.26.4
scikit-learn==1.4.2
opencv-python==4.9.0.80
Pillow==10.3.0
